[#3980] Position determined by index in collection.
As the resource position is a read only attribute, and can only be
changed with a separate reorder api call, we can't trust it when we're
saving the dataset and resource. Instead use a separate attribute
'template_position' which is defined as the index of the resource model
with the collection. This is always in sync with the visual order of
resources in the front end ui.