[#3977] Fix delete of unpersisted models.
If a backbone model has been created, but not persisted to the database,
attempts to destroy it will return `false` instead of a promise. Since
we expect a promise to chain a series of calls, we have to ensure a
destroyed model that returns `false` creates an immediately resolved
promise to work with.